+++ src/api-connection.service.ts
@@ -19,8 +19,6 @@ export class ApiConnectionService {
   private person_recover_token_url = 'https://api.wickeyappstore.com/person/recovery/token/';
   private person_recover_verify_url = 'https://api.wickeyappstore.com/person/recovery/verify/';
   private person_auth_url = 'https://api.wickeyappstore.com/person/auth/';
-  private app_url = 'https://api.wickeyappstore.com/apps/';
-  private featured_url = 'https://api.wickeyappstore.com/apps/featured/';
   private purchases_url = 'https://api.wickeyappstore.com/purchases/';
   private reviews_url = 'https://api.wickeyappstore.com/reviews/';
   private wasstore_url = 'https://api.wickeyappstore.com/wasstore/';
@@ -177,35 +175,6 @@ export class ApiConnectionService {
       }
     }
     return q_str.join('&');
-  }
-
-  // Returns app store apps {name: string, category: number, ordering: number}
-  getApps(_params?: any): Observable<[any]> {
-    this.handleHeaders();
-    // NOTE: Use share to avoid duplicate calls
-    const _query_string = this.encode_query_string(_params);
-    console.log('WASAPI: getApps', _query_string);
-    return this.http.get(`${this.app_url}?${_query_string}`, { headers: this.apiHeaders })
-      .map((res: any) => {
-        return this.extractData(res).apps;
-      }).catch(this.handleError).share();
-  }
-
-  /**
-   * This returns a list of featured groups of apps.
-   *
-   * @param [_params] None needed, just future proofing.
-   * @returns apps object {"groups": [{"id": number, "title": string, "created_time": number, "apps": [_app, _app, ...]}, {}, ...]}
-   */
-  getFeaturedGroups(_params?: any): Observable<[any]> {
-    this.handleHeaders();
-    // NOTE: Use share to avoid duplicate calls
-    const _query_string = this.encode_query_string(_params);
-    console.log('WASAPI: getFeaturedGroups', _query_string);
-    return this.http.get(`${this.featured_url}?${_query_string}`, { headers: this.apiHeaders })
-      .map((res: any) => {
-        return this.extractData(res).groups;
-      }).catch(this.handleError).share();
   }
 
   // Creates or updates person, returns person info